refactor: 인증 관련 로직은 모두 auth 폴더에 위치하도록 패키지 구조 변경
