+++ src/main/java/wad/seoul_nolgoat/auth/oauth2/CustomSuccessHandler.java
@@ -1,4 +1,4 @@
-package wad.seoul_nolgoat.auth;
+package wad.seoul_nolgoat.auth.oauth2;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
-import wad.seoul_nolgoat.auth.dto.OAuth2UserImpl;
+import wad.seoul_nolgoat.auth.oauth2.dto.OAuth2UserImpl;
 import wad.seoul_nolgoat.auth.service.AuthService;
 
 import java.io.IOException;
