+++ src/main/java/wad/seoul_nolgoat/auth/oauth2/dto/OAuth2Response.java
@@ -1,4 +1,4 @@
-package wad.seoul_nolgoat.auth.dto;
+package wad.seoul_nolgoat.auth.oauth2.dto;
 
 public interface OAuth2Response {
