+++ src/main/java/wad/seoul_nolgoat/auth/oauth2/dto/OAuth2UserImpl.java
@@ -1,4 +1,4 @@
-package wad.seoul_nolgoat.auth.dto;
+package wad.seoul_nolgoat.auth.oauth2.dto;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
