+++ src/main/java/wad/seoul_nolgoat/auth/web/dto/response/UserProfileDto.java
@@ -1,4 +1,4 @@
-package wad.seoul_nolgoat.web.auth.dto.response;
+package wad.seoul_nolgoat.auth.web.dto.response;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
