+++ src/main/java/wad/seoul_nolgoat/util/mapper/UserMapper.java
@@ -1,7 +1,7 @@
 package wad.seoul_nolgoat.util.mapper;
 
+import wad.seoul_nolgoat.auth.web.dto.response.UserProfileDto;
 import wad.seoul_nolgoat.domain.user.User;
-import wad.seoul_nolgoat.web.auth.dto.response.UserProfileDto;
 import wad.seoul_nolgoat.web.user.dto.request.UserSaveDto;
 import wad.seoul_nolgoat.web.user.dto.response.UserDetailsDto;
